refactor: move WalletChangeset to wallet module
